requirements.txt
discord.py==2.7.1
requests==2.31.0